requirements.txt
discord.py
python-a2s